[Add] : Debugging Example done

+++ cppTuto/Debugging Ex/Exercise_4/Knight.h
@@ -13,7 +13,7 @@ public:
 	bool IsDead();
 
 public:
-	int _hp;
+	long long int _hp;
 	int _attack;
 };
